% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euro.R
\name{euro}
\alias{euro}
\title{Formats number in EUR currency}
\usage{
euro(x)
}
\arguments{
\item{x}{number}
}
\value{
string
}
\description{
Formats number in EUR currency
}
\examples{
euro(1000)
euro(10.3241245125125)
}
