% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Look up the current price of a Bitcoin in USD}
\usage{
get_bitcoin_price(...)
}
\arguments{
\item{retried}{number of times the function already failed}
}
\value{
number
}
\description{
Look up the current price of a Bitcoin in USD
}
