% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_usdeur}
\alias{get_usdeur}
\title{Look up the value of a US Dollar in EURs}
\usage{
get_usdeur(retried = 0)
}
\arguments{
\item{retried}{number of times the function already failed}
}
\value{
number
}
\description{
Look up the value of a US Dollar in EURs
}
