% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_usdeurs}
\alias{get_usdeurs}
\title{Look up the historical time series of USD/EUR exchange rate}
\usage{
get_usdeurs(...)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}
}
\value{
data.table object with data and value columns
}
\description{
Look up the historical time series of USD/EUR exchange rate
}
