% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_usdeurs}
\alias{get_usdeurs}
\title{Look up the value of a US Dollar in Euro}
\usage{
get_usdeurs(...)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}
}
\value{
\code{data.table} object with dates and values
}
\description{
Look up the value of a US Dollar in Euro
}
